+++ crates/utilities/traits.rs
@@ -7,7 +7,7 @@ pub trait Object {
     /// вершины в локальных координатах
     fn vertices(&self) -> &[Vertex];
     /// индексы (триангулированные, u16)
-    fn indices(&self) -> &[u16];
+    fn indices(&self) -> &[u32];
     /// модельная матрица (локальная -> мировая)
     fn model_matrix(&self) -> Mat4;
